requirements_dev.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
psycopg2-binary>=2.9.9
pytest-mock>=3.14.0
pytest-postgresql>=6.0.0